ui changed

+++ lib/app/controllers/download_controller.dart
@@ -11,58 +11,74 @@ class DownloadController extends GetxController {
   final YoutubeExplode _yt = YoutubeExplode();
   final List<FileData> _files = [];
   List<FileData> get files => _files;
-
+  bool get isDownloading =>_isDownloading;
+  bool  _isDownloading =false;
   @override
   void onInit() {
     super.onInit();
     loadStoredFiles();
   }
 
-  void downloadUrl(String url) {
-    if (isPlaylist(url)) {
-      getPlayListVideoIds(url);
-    } else {
-      getVideoIdFromUrl(url);
-    }
+
+
+  /// Initializes the download process based on URL type (Playlist/Single Video)
+  void initUrl(String url) {
+    isPlaylist(url) ? _fetchPlaylistVideos(url) : _fetchVideoDetails(url);
   }
 
   bool isPlaylist(String url) => url.contains("playlist?list=");
 
-  Future<void> getPlayListVideoIds(String url) async {
+  bool _isFileExists(String id) => _files.any((file) => file.id == id);
+
+  /// Fetches and downloads videos from a playlist
+  Future<void> _fetchPlaylistVideos(String url) async {
     try {
-      String playlistId = extractPlaylistId(url);
+      String playlistId = _extractPlaylistId(url);
       await for (var video in _yt.playlists.getVideos(playlistId)) {
-        _files.add(FileData(id: video.id.value, name: video.title));
-        update();
-        download(video.id.value);
+        if (!_isFileExists(video.id.value)) {
+          _addFile(video);
+        }
       }
-      saveFileData();
+      await _saveFileData();
+      await  downloadAll();
+      _isDownloading=false;
     } catch (e) {
       Snack.showErrorMessage(e.toString());
     }
   }
 
-  void updateFileData(String id, {bool? isDownload, String? path, double? downloadProgress}) {
-    for (int i = 0; i < _files.length; i++) {
+ Future downloadAll() async {
+   if (isDownloading) return; // Prevent multiple calls
+    for (var file in _files) {
+      if(!file.isDownload){
+       await downloadFile(file.id);
+      }
+    }
+  }
+
+  /// Updates file details
+  void _updateFileData(String id, {bool? isDownload, String? path, double? downloadProgress}) {
+    for (var i = 0; i < _files.length; i++) {
       if (_files[i].id == id) {
-        _files[i] = _files[i].copyWith(isDownload: isDownload, path: path, downloadProgress: downloadProgress);
+        _files[i] = _files[i].copyWith(
+          isDownload: isDownload ?? _files[i].isDownload,
+          path: path ?? _files[i].path,
+          downloadProgress: downloadProgress ?? _files[i].downloadProgress,
+        );
       }
     }
     update();
-    saveFileData();
+    _saveFileData();
   }
 
   Future<void> loadStoredFiles() async {
     try {
       final directory = await getApplicationDocumentsDirectory();
       final file = File('${directory.path}/downloads.json');
-
       if (await file.exists()) {
         String contents = await file.readAsString();
-        List<dynamic> jsonData = jsonDecode(contents);
-
         _files.clear();
-        _files.addAll(jsonData.map((e) => FileData.fromJson(e)));
+        _files.addAll(List<FileData>.from(jsonDecode(contents).map((e) => FileData.fromJson(e))));
         update();
       }
     } catch (e) {
@@ -70,39 +86,32 @@ class DownloadController extends GetxController {
     }
   }
 
-  Future<void> saveFileData() async {
+  Future<void> _saveFileData() async {
     try {
       final directory = await getApplicationDocumentsDirectory();
       final file = File('${directory.path}/downloads.json');
-
-      String jsonData = jsonEncode(_files.map((e) => e.toJson()).toList());
-      await file.writeAsString(jsonData);
+      await file.writeAsString(jsonEncode(_files.map((e) => e.toJson()).toList()));
     } catch (e) {
       print("Error saving file data: $e");
     }
   }
 
-  String extractPlaylistId(String url) {
+  String _extractPlaylistId(String url) {
     var parts = url.split("list=");
-    if (parts.length > 1) {
-      return parts[1].split('&')[0];
-    }
+    if (parts.length > 1) return parts[1].split('&')[0];
     throw 'Invalid playlist URL';
   }
 
-  Future<void> getVideoIdFromUrl(String url) async {
+  Future<void> _fetchVideoDetails(String url) async {
     try {
       final RegExp videoIdPattern = RegExp(r'(?:v=|\/)([0-9A-Za-z_-]{11}).*');
       final match = videoIdPattern.firstMatch(url);
-
       if (match != null && match.groupCount >= 1) {
         final videoId = match.group(1);
         if (videoId != null) {
-          final video = await _yt.videos.get(videoId);
-          _files.add(FileData(id: video.id.value, name: video.title));
-          update();
-          await download(videoId);
-          saveFileData();
+          var video = await _yt.videos.get(videoId);
+          _addFile(video);
+          await downloadFile(videoId);
         } else {
           throw 'Invalid video URL';
         }
@@ -114,65 +123,59 @@ class DownloadController extends GetxController {
     }
   }
 
-  Future<void> download(String videoId) async {
+  void _addFile(Video video) {
+    final file = FileData(
+      id: video.id.value,
+      name: video.title,
+      description: video.description,
+      thumbnails: video.thumbnails.mediumResUrl,
+      duration: video.duration,
+      publishDate: video.publishDate,
+    );
+    _files.add(file);
+    _saveFileData();
+    update();
+  }
+
+  Future<void> downloadFile(String videoId) async {
     try {
+      print("DOWNLOADING....");
+      _isDownloading=true;
       var video = await _yt.videos.get(videoId);
       var manifest = await _yt.videos.streams.getManifest(video.id);
       var audioStreamInfo = manifest.audioOnly.withHighestBitrate();
-
-      if (audioStreamInfo == null) {
-        throw "No audio stream available for this video.";
-      }
-
+      if (audioStreamInfo == null) throw "No audio stream available.";
       var audioStream = _yt.videos.streamsClient.get(audioStreamInfo);
       var tempDir = await getApplicationDocumentsDirectory();
       var audioFilePath = '${tempDir.path}/${video.id}.webm';
-      var audioFile = File(audioFilePath);
-      var audioFileStream = audioFile.openWrite();
-
+      var audioFile = File(audioFilePath).openWrite();
       int totalBytes = audioStreamInfo.size.totalBytes;
       int downloadedBytes = 0;
-
       await for (var chunk in audioStream) {
-        audioFileStream.add(chunk);
+        audioFile.add(chunk);
         downloadedBytes += chunk.length;
-
-        double progress = downloadedBytes / totalBytes;
-        updateFileData(videoId, downloadProgress: progress);
-        update();
+        _updateFileData(videoId, downloadProgress: downloadedBytes / totalBytes);
       }
-
-      await audioFileStream.flush();
-      await audioFileStream.close();
-
-      updateFileData(videoId, isDownload: true, path: audioFile.path, downloadProgress: 1.0);
-      update();
+      await audioFile.flush();
+      await audioFile.close();
+      _updateFileData(videoId, isDownload: true, path: audioFilePath, downloadProgress: 1.0);
     } catch (e) {
-      print('Error downloading and saving video: $e');
+      print('Error downloading file: $e');
     }
   }
 
-  /// **New Method: Delete a Downloaded File**
   Future<void> deleteFile(String id) async {
     try {
-      // Find the file in the list
       final fileIndex = _files.indexWhere((file) => file.id == id);
       if (fileIndex == -1) return;
-
       final filePath = _files[fileIndex].path;
       if (filePath != null && filePath.isNotEmpty) {
         final file = File(filePath);
-        if (await file.exists()) {
-          await file.delete(); // Delete the actual file
-        }
+        if (await file.exists()) await file.delete();
       }
-
-      // Remove file from list
       _files.removeAt(fileIndex);
       update();
-      saveFileData();
-
-      print("File deleted successfully.");
+      _saveFileData();
     } catch (e) {
       print("Error deleting file: $e");
     }
